requirements.txt
asyncio==3.4.3
beautifulsoup4==4.9.3
discord.py==1.5.1
requests==2.25.1
youtube-search==1.1.1